feat(ts): add platform-specific #file_tool_modifier to avoid using "node:crypto" in Cloudflare Workers runtime
