+++ ts/packages/core/src/utils/modifiers/FileToolModifier.node.ts
@@ -1,139 +1,19 @@
 import {
-  JSONSchemaProperty,
   Tool,
   ToolExecuteParams,
   ToolExecuteResponse,
+  JSONSchemaProperty,
 } from '../../types/tool.types';
 import ComposioClient from '@composio/client';
 import logger from '../logger';
 import { ComposioFileUploadError } from '../../errors/FileModifierErrors';
 import { downloadFileFromS3, getFileDataAfterUploadingToS3 } from '../fileUtils';
-
-/**
- * Transforms a single JSON schema property, recursively handling nested properties,
- * anyOf, oneOf, and allOf.
- */
-const transformSchema = (property: JSONSchemaProperty): JSONSchemaProperty => {
-  if (property.file_uploadable) {
-    // Transform file-uploadable property
-    return {
-      title: property.title,
-      description: property.description,
-      format: 'path',
-      type: 'string',
-      file_uploadable: true,
-    };
-  }
-
-  const newProperty = { ...property };
-
-  if (property.type === 'object' && property.properties) {
-    // Recursively transform nested properties
-    newProperty.properties = transformProperties(property.properties);
-  }
-
-  if (property.anyOf) {
-    newProperty.anyOf = property.anyOf.map(transformSchema);
-  }
-
-  if (property.oneOf) {
-    newProperty.oneOf = property.oneOf.map(transformSchema);
-  }
-
-  if (property.allOf) {
-    newProperty.allOf = property.allOf.map(transformSchema);
-  }
-
-  if (property.items) {
-    if (Array.isArray(property.items)) {
-      newProperty.items = property.items.map(transformSchema);
-    } else {
-      newProperty.items = transformSchema(property.items);
-    }
-  }
-
-  return newProperty;
-};
-
-/**
- * Transforms the properties of the tool schema to include the file upload URL.
- *
- * Attaches the format: 'path' to the properties that are file uploadable for agents.
- *
- * @param properties - The properties of the tool schema.
- * @returns The transformed properties.
- */
-const transformProperties = (properties: JSONSchemaProperty): JSONSchemaProperty => {
-  const newProperties: JSONSchemaProperty = {};
-
-  for (const [key, property] of Object.entries(properties) as [string, JSONSchemaProperty][]) {
-    newProperties[key] = transformSchema(property);
-  }
-
-  return newProperties;
-};
-
-/**
- * Recursively checks if a schema (or any of its variants) contains a specific file property.
- */
-const schemaHasFileProperty = (
-  schema: JSONSchemaProperty | undefined,
-  property: 'file_uploadable' | 'file_downloadable'
-): boolean => {
-  if (!schema) return false;
-  if (schema[property]) return true;
-
-  // Check nested properties
-  if (schema.properties) {
-    for (const prop of Object.values(schema.properties) as JSONSchemaProperty[]) {
-      if (schemaHasFileProperty(prop, property)) return true;
-    }
-  }
-
-  // Check anyOf/oneOf/allOf variants
-  if (schema.anyOf) {
-    for (const variant of schema.anyOf) {
-      if (schemaHasFileProperty(variant, property)) return true;
-    }
-  }
-  if (schema.oneOf) {
-    for (const variant of schema.oneOf) {
-      if (schemaHasFileProperty(variant, property)) return true;
-    }
-  }
-  if (schema.allOf) {
-    for (const variant of schema.allOf) {
-      if (schemaHasFileProperty(variant, property)) return true;
-    }
-  }
-
-  // Check array items
-  if (schema.items) {
-    if (Array.isArray(schema.items)) {
-      for (const item of schema.items) {
-        if (schemaHasFileProperty(item, property)) return true;
-      }
-    } else {
-      if (schemaHasFileProperty(schema.items, property)) return true;
-    }
-  }
-
-  return false;
-};
-
-/**
- * Recursively checks if a schema (or any of its variants) contains file_uploadable properties.
- */
-const schemaHasFileUploadable = (schema: JSONSchemaProperty | undefined): boolean => {
-  return schemaHasFileProperty(schema, 'file_uploadable');
-};
-
-/**
- * Recursively checks if a schema (or any of its variants) contains file_downloadable properties.
- */
-const schemaHasFileDownloadable = (schema: JSONSchemaProperty | undefined): boolean => {
-  return schemaHasFileProperty(schema, 'file_downloadable');
-};
+import {
+  isPlainObject,
+  transformProperties,
+  schemaHasFileUploadable,
+  schemaHasFileDownloadable,
+} from './FileToolModifier.utils.neutral';
 
 /**
  * Recursively walks a runtime value and its matching JSON-Schema node,
@@ -352,11 +232,6 @@ const hydrateDownloads = async (
   return value;
 };
 
-// Small helper to recognise plain objects
-function isPlainObject(val: unknown): val is Record<string, unknown> {
-  return typeof val === 'object' && val !== null && !Array.isArray(val);
-}
-
 export class FileToolModifier {
   private client: ComposioClient;
 
@@ -364,16 +239,6 @@ export class FileToolModifier {
     this.client = client;
   }
 
-  /**
-   * Modifies the tool schema to include the file upload URL.
-   *
-   * @description This modifier is used to upload a file to the Composio platform and replace the file path with the file upload URL.
-   *
-   * @param _toolSlug - The slug of the tool that is being executed.
-   * @param _toolkitSlug - The slug of the toolkit that is being executed.
-   * @param schema - The schema of the tool.
-   * @returns The schema with the file upload URL included.
-   */
   async modifyToolSchema(toolSlug: string, toolkitSlug: string, schema: Tool): Promise<Tool> {
     if (!schema.inputParameters?.properties) {
       return schema;
@@ -390,15 +255,6 @@ export class FileToolModifier {
     };
   }
 
-  /**
-   * Modifies the input parameters to include the file upload URL.
-   *
-   * @description This modifier is used to upload a file to the Composio platform and replace the file path with the file upload URL.
-   *
-   * @param toolSlug - The slug of the tool that is being executed.
-   * @param toolkitSlug - The slug of the toolkit that is being executed.
-   *
-   */
   async fileUploadModifier(
     tool: Tool,
     options: {
@@ -427,17 +283,6 @@ export class FileToolModifier {
     }
   }
 
-  /**
-   * Modifies the result to include the file download URL.
-   *
-   * @description This modifier is used to download a file and
-   *
-   * @param tool - The tool schema containing output parameters.
-   * @param toolSlug - The slug of the tool that is being executed.
-   * @param toolkitSlug - The slug of the toolkit that is being executed.
-   * @param result - The result of the tool execution.
-   * @returns The result with the file download URL included.
-   */
   async fileDownloadModifier(
     tool: Tool,
     options: {
