+++ ts/packages/core/tsdown.config.ts
@@ -7,10 +7,18 @@ export default defineConfig({
   copy: [{ from: 'pack/generated/*', to: '.', flatten: false }],
   entry: [
     'src/index.ts',
+
+    // #platform
     'src/platform/node.ts',
     'src/platform/workerd.ts',
+
+    // #files
     'src/models/Files.node.ts',
     'src/models/Files.workerd.ts',
+
+    // #file_tool_modifier
+    'src/utils/modifiers/FileToolModifier.node.ts',
+    'src/utils/modifiers/FileToolModifier.workerd.ts',
   ],
   attw: {
     ...baseConfig.attw,
@@ -33,5 +41,5 @@ export default defineConfig({
    * We don't want to accidentally bundle `node:*` packages (e.g., `node:module`)
    * as not all of them are available in Cloudflare Workers / Vercel Edge runtimes.
    */
-  external: [...(baseConfig.external ?? []), '#platform', '#files'],
+  external: [...(baseConfig.external ?? []), '#platform', '#files', '#file_tool_modifier'],
 });
