+++ ts/packages/core/vitest.config.ts
@@ -9,6 +9,10 @@ export default defineConfig({
     alias: {
       '#platform': path.resolve(__dirname, 'src/platform/node.ts'),
       '#files': path.resolve(__dirname, 'src/models/Files.node.ts'),
+      '#file_tool_modifier': path.resolve(
+        __dirname,
+        'src/utils/modifiers/FileToolModifier.node.ts'
+      ),
     },
   },
   test: {
